Break the requirementsOperations import cycle
requirementsOperations.ts owned both the CLI log helpers and the
executable-path constants, so every other requirements module imported
from it — and operations needed everything else back. The compromise
was lazy require() calls scattered through generateRequirements,
generateTestsFromRequirements, importRequirements, exportRequirements,
offerTraceabilityInferenceIfMissing, and setupReqs2XExecutablePaths,
which masked the cycle rather than fixing it.

Extract two leaf modules:

- requirementsLog.ts holds the CLI output channel + logCliOperation /
  logCliError. Imported by everyone (requirementsExecutables,
  processRunner, requirementsOperations) without any back-edges.
- requirementsExecutables.ts owns the four executable-path module
  bindings (CODE2REQS / REQS2TESTS / PANREQ / LLM2CHECK), the
  setupReqs2XExecutablePaths function, and getAutoreqExecutableDirectory
  (moved out of availability.ts where it didn't really belong).

After the move:
  rgwPath, requirementsLog              (leaves)
  requirementsExecutables, availability  (depend on the above)
  llmProvider, processRunner             (depend on executables / log)
  rgwIo                                  (depends on processRunner +
                                          executables + rgwPath)
  requirementsOperations                 (top, depends on all)

All require() calls inside src/requirements/ are gone; every module
imports at the top of the file. Behaviour is unchanged.
